requirements.txt
selenium==3.141.0
pytest==6.2.4
mimesis
allure-pytest